Replace paginated design with a single mono-page
Adds a "position" property to each page that controls order in the
mono-page.
Adds a custom Eleventy Collection called `sortedPages`(.eleventy.js)
that is sorted by the `position` in each page.

Replaces `index.md` with a new `index.html` that uses Liquid to loop
over the new custom Collection and render each page in order.